requirements.txt
transformers>=4.41.0
datasets>=2.19.0
peft>=0.11.1
accelerate>=0.30.1
python-dotenv>=1.0.1
torch>=2.2.2

